feat: serialize avatar_url for room list
