+++ guest-js/stores/rooms-collection.svelte.ts
@@ -39,7 +39,9 @@ export class RoomsCollection extends RuneStore<RoomsCollectionType> {
             roomName: "default",
             numUnreadMentions: 0,
             numUnreadMessages: 0,
+            tags: null,
             latest: null,
+            avatar: null,
             altAlias: null,
             canonicalAlias: null,
             hasBeenPaginated: false,
